changes in Response

+++ src/main/java/com/easy/request/SaveFormRequestDTO.java
@@ -40,7 +40,7 @@ public class SaveFormRequestDTO {
     private Long noOfBalconies;
 
     // House
-    public Integer numberOfFloors;
+    public Long numberOfFloors;
     public Double builtUpArea;
     public Integer constructionYear;
     private Long permittedFloors;
